Moving all procore related apis to integrations api class [CAP-528]
